refactor(build-cli): remove barrel files (AB#8118)
Delete 4 barrel index.ts files from build-cli and update all 55
importers to use direct module imports instead. Barrel files removed:
- src/library/commands/index.ts
- src/handlers/index.ts
- src/machines/index.ts
- src/library/index.ts

Kept src/library/repoPolicyCheck/index.ts which has real aggregation
logic (builds policyHandlers array), not a pure re-export barrel.
