+++ build-tools/packages/build-cli/src/BasePackageCommand.ts
@@ -17,7 +17,7 @@ import {
 	selectAndFilterPackages,
 } from "./filter.js";
 import { filterFlags, type PackageSelectionDefault, selectionFlags } from "./flags.js";
-import { BaseCommand } from "./library/index.js";
+import { BaseCommand } from "./library/commands/base.js";
 
 /**
  * Commands that run operations per project.
