+++ build-tools/packages/build-cli/src/commands/check/changeset.ts
@@ -7,7 +7,7 @@ import { Flags } from "@oclif/core";
 import chalk from "picocolors";
 import { sortPackageJson as sortJson } from "sort-package-json";
 
-import { BaseCommand } from "../../library/index.js";
+import { BaseCommand } from "../../library/commands/base.js";
 
 export default class CheckChangesetCommand extends BaseCommand<typeof CheckChangesetCommand> {
 	static readonly summary =
