+++ build-tools/packages/build-cli/src/commands/check/layers.ts
@@ -8,7 +8,8 @@ import path from "node:path";
 
 import { Flags } from "@oclif/core";
 
-import { BaseCommand, LayerGraph } from "../../library/index.js";
+import { BaseCommand } from "../../library/commands/base.js";
+import { LayerGraph } from "../../library/layerGraph.js";
 // eslint-disable-next-line import-x/no-internal-modules -- AB#8118 tracks removing the barrel files and importing directly from the submodules, including disabling this rule.
 import { writeFileWithLineFeeds } from "../../library/text.js";
