+++ build-tools/packages/build-cli/src/commands/check/prApproval.ts
@@ -6,7 +6,7 @@
 import { Flags } from "@oclif/core";
 
 import { githubActionsFlag, githubTokenFlag } from "../../flags.js";
-
+import { BaseCommand } from "../../library/commands/base.js";
 import {
 	type GitHubProps,
 	isPrApprovedByTeam,
@@ -14,7 +14,6 @@ import {
 	// We are slowly moving away from barrel files and these APIs are only used here.
 	// eslint-disable-next-line import-x/no-internal-modules
 } from "../../library/githubRest.js";
-import { BaseCommand } from "../../library/index.js";
 
 /**
  * This command class is used to merge two branches based on the batch size provided.
