+++ build-tools/packages/build-cli/src/commands/generate/changeset.ts
@@ -15,11 +15,9 @@ import chalk from "picocolors";
 import prompts from "prompts";
 
 import { releaseGroupFlag } from "../../flags.js";
-import {
-	BaseCommand,
-	type FluidCustomChangesetMetadata,
-	getDefaultBumpTypeForBranch,
-} from "../../library/index.js";
+import { getDefaultBumpTypeForBranch } from "../../library/branches.js";
+import type { FluidCustomChangesetMetadata } from "../../library/changesets.js";
+import { BaseCommand } from "../../library/commands/base.js";
 
 /**
  * If more than this number of packages are changed relative to the selected branch, the user will be prompted to select
