+++ build-tools/packages/build-cli/src/commands/generate/releaseNotes.ts
@@ -17,18 +17,18 @@ import remarkToc from "remark-toc";
 import type { ReleaseNotesSection } from "../../config.js";
 import { releaseGroupFlag } from "../../flags.js";
 import {
-	BaseCommand,
 	DEFAULT_CHANGESET_PATH,
-	difference,
 	fluidCustomChangeSetMetadataDefaults,
 	groupBySection,
 	loadChangesets,
 	UNKNOWN_SECTION,
-} from "../../library/index.js";
+} from "../../library/changesets.js";
+import { BaseCommand } from "../../library/commands/base.js";
 // eslint-disable-next-line import-x/no-internal-modules
 import { addHeadingLinks, stripSoftBreaks } from "../../library/markdown.js";
 // eslint-disable-next-line import-x/no-internal-modules
 import { RELEASE_NOTES_TOC_LINK_TEXT } from "../../library/releaseNotes.js";
+import { difference } from "../../library/sets.js";
 
 /**
  * Generates release notes from individual changeset files.
