+++ build-tools/packages/build-cli/src/commands/info.ts
@@ -8,8 +8,9 @@ import { Flags } from "@oclif/core";
 import * as changeCase from "change-case";
 import { type ColumnUserConfig, table } from "table";
 import { releaseGroupFlag } from "../flags.js";
+import { BaseCommand } from "../library/commands/base.js";
 // eslint-disable-next-line import-x/no-deprecated
-import { BaseCommand, isMonoRepoKind } from "../library/index.js";
+import { isMonoRepoKind } from "../library/context.js";
 
 /**
  * Each displayed column in the report is generated by a corresponding
