+++ build-tools/packages/build-cli/src/commands/modify/fluid-imports.ts
@@ -10,7 +10,9 @@ import { readFile } from "node:fs/promises";
 import { Flags } from "@oclif/core";
 import * as JSON5 from "json5";
 import { type ImportDeclaration, ModuleKind, Project, type SourceFile } from "ts-morph";
-import { ApiLevel, BaseCommand, getApiExports, isKnownApiLevel } from "../../library/index.js";
+import { ApiLevel, isKnownApiLevel } from "../../library/apiLevel.js";
+import { BaseCommand } from "../../library/commands/base.js";
+import { getApiExports } from "../../library/typescriptApi.js";
 import type { CommandLogger } from "../../logging.js";
 
 const maxConcurrency = 4;
