+++ build-tools/packages/build-cli/src/commands/promote/package.ts
@@ -5,7 +5,8 @@
 
 import { Flags } from "@oclif/core";
 import { semverFlag } from "../../flags.js";
-import { BaseCommand, readLines } from "../../library/index.js";
+import { BaseCommand } from "../../library/commands/base.js";
+import { readLines } from "../../library/text.js";
 
 interface PackagePromotionErrorResponse {
 	success: string;
