+++ build-tools/packages/build-cli/src/commands/publish/bundleStats.ts
@@ -8,7 +8,7 @@ import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { Flags } from "@oclif/core";
 
-import { BaseCommand } from "../../library/index.js";
+import { BaseCommand } from "../../library/commands/base.js";
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
