+++ build-tools/packages/build-cli/src/commands/release.ts
@@ -24,12 +24,13 @@ import {
 import {
 	FluidReleaseStateHandler,
 	type FluidReleaseStateHandlerData,
-	type StateHandler,
-} from "../handlers/index.js";
+} from "../handlers/fluidReleaseStateHandler.js";
+import type { StateHandler } from "../handlers/stateHandlers.js";
 import { PromptWriter } from "../instructionalPromptWriter.js";
+import { getDefaultBumpTypeForBranch } from "../library/branches.js";
 // eslint-disable-next-line import-x/no-deprecated
-import { getDefaultBumpTypeForBranch, MonoRepoKind } from "../library/index.js";
-import { FluidReleaseMachine } from "../machines/index.js";
+import { MonoRepoKind } from "../library/context.js";
+import { FluidReleaseMachine } from "../machines/fluidReleaseMachine.js";
 import { getRunPolicyCheckDefault } from "../repoConfig.js";
 import { StateMachineCommand } from "../stateMachineCommand.js";
