+++ build-tools/packages/build-cli/src/commands/release/fromTag.ts
@@ -15,7 +15,8 @@ import { sortPackageJson as sortJson } from "sort-package-json";
 
 import { findPackageOrReleaseGroup } from "../../args.js";
 // eslint-disable-next-line import-x/no-deprecated
-import { MonoRepoKind, sortVersions } from "../../library/index.js";
+import { MonoRepoKind } from "../../library/context.js";
+import { sortVersions } from "../../library/package.js";
 import type { ReleaseGroup, ReleasePackage } from "../../releaseGroups.js";
 import { ReleaseReportBaseCommand, type ReleaseSelectionMode } from "./report.js";
