+++ build-tools/packages/build-cli/src/commands/release/report-unreleased.ts
@@ -11,7 +11,8 @@ import { Flags } from "@oclif/core";
 import { formatISO } from "date-fns";
 
 import { semverFlag } from "../../flags.js";
-import { BaseCommand, type ReleaseReport, toReportKind } from "../../library/index.js";
+import { BaseCommand } from "../../library/commands/base.js";
+import { type ReleaseReport, toReportKind } from "../../library/release.js";
 
 export class UnreleasedReportCommand extends BaseCommand<typeof UnreleasedReportCommand> {
 	static readonly summary =
