+++ build-tools/packages/build-cli/src/commands/rename-types.ts
@@ -5,7 +5,7 @@
 
 import * as fs from "node:fs";
 import * as path from "node:path";
-import { BaseCommand } from "../library/index.js";
+import { BaseCommand } from "../library/commands/base.js";
 
 function renameFilesInDir(dir: string, extension: string, newExtension: string): void {
 	const files = fs.readdirSync(dir);
