+++ build-tools/packages/build-cli/src/commands/report/codeCoverage.ts
@@ -11,13 +11,13 @@ import {
 } from "../../codeCoverage/compareCodeCoverage.js";
 import { getCommentForCodeCoverageDiff } from "../../codeCoverage/getCommentForCodeCoverage.js";
 import type { IAzureDevopsBuildCoverageConstants } from "../../library/azureDevops/constants.js";
+import { BaseCommand } from "../../library/commands/base.js";
 import {
 	createOrUpdateCommentOnPr,
 	type GitHubProps,
 	getChangedFilePaths,
 	getCommentBody,
 } from "../../library/githubRest.js";
-import { BaseCommand } from "../../library/index.js";
 
 // Unique identifier for the comment made on the PR. This is used to identify the comment
 // and update it based on a new build.
