+++ build-tools/packages/build-cli/src/commands/transform/releaseNotes.ts
@@ -12,7 +12,7 @@ import remarkGithub, { defaultBuildUrl } from "remark-github";
 import admonitions from "remark-github-beta-blockquote-admonitions";
 import remarkToc from "remark-toc";
 
-import { BaseCommand } from "../../library/index.js";
+import { BaseCommand } from "../../library/commands/base.js";
 import {
 	addHeadingLinks,
 	removeHeadingsAtLevel,
