+++ build-tools/packages/build-cli/src/filter.ts
@@ -6,7 +6,7 @@
 import path from "node:path";
 import { type MonoRepo, Package } from "@fluidframework/build-tools";
 import type { filterFlags, PackageSelectionDefault, selectionFlags } from "./flags.js";
-import type { Context } from "./library/index.js";
+import type { Context } from "./library/context.js";
 import { knownReleaseGroups, type ReleaseGroup } from "./releaseGroups.js";
 
 /**
