+++ build-tools/packages/build-cli/src/flags.ts
@@ -7,7 +7,7 @@ import { Flags } from "@oclif/core";
 import * as semver from "semver";
 
 // eslint-disable-next-line import-x/no-deprecated
-import { MonoRepoKind } from "./library/index.js";
+import { MonoRepoKind } from "./library/context.js";
 
 /**
  * An iterator that returns only the Enum values of MonoRepoKind.
@@ -29,7 +29,7 @@ import {
 	type VersionBumpType,
 	type VersionScheme,
 } from "@fluid-tools/version-tools";
-import type { DependencyUpdateType } from "./library/index.js";
+import type { DependencyUpdateType } from "./library/bump.js";
 import { isReleaseGroup, type ReleaseGroup } from "./releaseGroups.js";
 
 /**
