--- build-tools/packages/build-cli/src/handlers/index.ts
@@ -1,11 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export {
-	FluidReleaseStateHandler,
-	type FluidReleaseStateHandlerData,
-} from "./fluidReleaseStateHandler.js";
-export { InitFailedStateHandler } from "./initFailedStateHandler.js";
-export { BaseStateHandler, type StateHandler } from "./stateHandlers.js";
