+++ build-tools/packages/build-cli/src/handlers/initFailedStateHandler.ts
@@ -6,7 +6,7 @@
 import type { Machine } from "jssm";
 
 import type { CommandLogger } from "../logging.js";
-import type { MachineState } from "../machines/index.js";
+import type { MachineState } from "../machines/types.js";
 import { BaseStateHandler } from "./stateHandlers.js";
 
 /**
