+++ build-tools/packages/build-cli/src/handlers/promptFunctions.ts
@@ -8,13 +8,13 @@ import chalk from "picocolors";
 
 import { type InstructionalPrompt, mapADOLinks } from "../instructionalPromptWriter.js";
 import {
-	difference,
 	generateReleaseBranchName,
 	getDefaultBumpTypeForBranch,
-	getPreReleaseDependencies,
-} from "../library/index.js";
+} from "../library/branches.js";
+import { getPreReleaseDependencies } from "../library/package.js";
+import { difference } from "../library/sets.js";
 import type { CommandLogger } from "../logging.js";
-import type { MachineState } from "../machines/index.js";
+import type { MachineState } from "../machines/types.js";
 import { isReleaseGroup } from "../releaseGroups.js";
 import type { FluidReleaseStateHandlerData } from "./fluidReleaseStateHandler.js";
 import type { StateHandlerFunction } from "./stateHandlers.js";
