+++ build-tools/packages/build-cli/src/instructionalPromptWriter.ts
@@ -7,7 +7,8 @@ import { StringBuilder } from "@rushstack/node-core-library";
 import chalk from "picocolors";
 
 // eslint-disable-next-line import-x/no-deprecated
-import { indentString, MonoRepoKind } from "./library/index.js";
+import { MonoRepoKind } from "./library/context.js";
+import { indentString } from "./library/text.js";
 import type { CommandLogger } from "./logging.js";
 import type { ReleaseGroup, ReleasePackage } from "./releaseGroups.js";
