--- build-tools/packages/build-cli/src/library/commands/index.ts
@@ -1,11 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { BaseCommand, BaseCommandWithBuildProject } from "./base.js";
-export { unscopedPackageNameString } from "./constants.js";
-export {
-	GenerateEntrypointsCommand,
-	getGenerateEntrypointsOutput,
-} from "./generateEntrypoints.js";
