--- build-tools/packages/build-cli/src/library/index.ts
@@ -1,67 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { ApiLevel, isKnownApiLevel, knownApiLevels } from "./apiLevel.js";
-export {
-	createBumpBranch,
-	generateBumpDepsBranchName,
-	generateBumpDepsCommitMessage,
-	generateBumpVersionBranchName,
-	generateBumpVersionCommitMessage,
-	generateReleaseBranchName,
-	getDefaultBumpTypeForBranch,
-	getReleaseSourceForReleaseGroup,
-} from "./branches.js";
-export {
-	bumpReleaseGroup,
-	type DependencyUpdateType,
-	isDependencyUpdateType,
-} from "./bump.js";
-export {
-	DEFAULT_CHANGESET_PATH,
-	type FluidCustomChangesetMetadata,
-	fluidCustomChangeSetMetadataDefaults,
-	groupByMainPackage,
-	groupBySection,
-	loadChangesets,
-	UNKNOWN_SECTION,
-} from "./changesets.js";
-export {
-	BaseCommand,
-	BaseCommandWithBuildProject,
-	GenerateEntrypointsCommand,
-	unscopedPackageNameString,
-} from "./commands/index.js";
-export { Context, isMonoRepoKind, MonoRepoKind, type VersionDetails } from "./context.js";
-export { getDisplayDate, getDisplayDateRelative } from "./dates.js";
-export { getVersionsFromTags, Repository } from "./git.js";
-export { createPullRequest, getCommitInfo, pullRequestExists } from "./github.js";
-export { LayerGraph } from "./layerGraph.js";
-export {
-	ensureDevDependencyExists,
-	filterVersionsOlderThan,
-	generateReleaseGitTagName,
-	getFluidDependencies,
-	getPreReleaseDependencies,
-	getTarballName,
-	isReleased,
-	npmCheckUpdates,
-	type PackageVersionMap,
-	type PreReleaseDependencies,
-	setVersion,
-	sortVersions,
-} from "./package.js";
-export {
-	getRanges,
-	type ReleaseRanges,
-	type ReleaseReport,
-	type ReportKind,
-	toReportKind,
-} from "./release.js";
-export { ReleaseLevel } from "./releaseLevel.js";
-export { type Handler, policyHandlers } from "./repoPolicyCheck/index.js";
-export { difference } from "./sets.js";
-export { getIndent, indentString, readLines } from "./text.js";
-export { getApiExports } from "./typescriptApi.js";
