+++ build-tools/packages/build-cli/src/library/repoPolicyCheck/fluidBuildDatabase.ts
@@ -12,7 +12,7 @@ import path from "node:path";
 import { type Package, TscUtils } from "@fluidframework/build-tools";
 import type { TsConfigJson } from "type-fest";
 
-import { getGenerateEntrypointsOutput } from "../commands/index.js";
+import { getGenerateEntrypointsOutput } from "../commands/generateEntrypoints.js";
 
 type PackageName = string;
 type Script = string;
