--- build-tools/packages/build-cli/src/machines/index.ts
@@ -1,7 +1,0 @@
-/*!
- * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
- * Licensed under the MIT License.
- */
-
-export { FluidReleaseMachine } from "./fluidReleaseMachine.js";
-export type { MachineState } from "./types.js";
