+++ build-tools/packages/build-cli/src/test/commands/list.test.ts
@@ -8,7 +8,7 @@ import { expect } from "chai";
 import { describe, it } from "mocha";
 
 import type { PackageNamePolicyConfig } from "../../config.js";
-import { Context } from "../../library/index.js";
+import { Context } from "../../library/context.js";
 import {
 	type Feed,
 	feeds,
