+++ build-tools/packages/build-cli/src/test/commands/release.test.ts
@@ -7,7 +7,7 @@ import { runCommand } from "@oclif/test";
 import { expect } from "chai";
 import { describe, it } from "mocha";
 
-import { FluidReleaseMachine } from "../../machines/index.js";
+import { FluidReleaseMachine } from "../../machines/fluidReleaseMachine.js";
 
 const knownUnhandledStates: Set<string> = new Set([
 	// Known unhandled states can be added here temporarily during development.
