+++ build-tools/packages/build-cli/src/test/library/package.test.ts
@@ -9,7 +9,7 @@ import { assert } from "chai";
 import { parseJSON } from "date-fns";
 import { describe, it } from "mocha";
 
-import type { VersionDetails } from "../../library/index.js";
+import type { VersionDetails } from "../../library/context.js";
 import {
 	ensureDevDependencyExists,
 	generateReleaseGitTagName,
